requirements.txt
fastapi==0.70.0
uvicorn==0.15.0
aioredis==2.0.0
pyngrok==5.1.0
httpx==0.27
openai==0.28.1
databases
asyncpg 
sqlalchemy[asyncio]
python-dotenv
aiogram==2.23.1
pydantic==1.10.16
